0.1.9 what's new in 0.1.9
add pathMap as third argument for gulp-twoside plugin, so that the module name in browser can be renamed, the packages can be reorganized, see src/lib and gulpfile.coffee for examples.
the first parameter "basepath" of gulp-twoside plugin add __dirname automatically now.
path in twoside(path) can be a package name now.